# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manage OAuth-connected Gmail and Microsoft mailboxes, proxies, and synced messages.",
        "title": "DeezMails API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/accounts/{id}/emails": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "List synced emails",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search sender, subject, and preview",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Folder ID or name; defaults to all folders",
                        "name": "folder",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "receivedAt",
                            "subject",
                            "from",
                            "createdAt"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}/emails/{messageID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "Get an email in JSON or raw RFC 822 format",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider message ID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "json",
                            "raw"
                        ],
                        "type": "string",
                        "description": "Response format",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Email"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}/folders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List available mailbox folders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Folder"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Email": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "body": {
                    "type": "string"
                },
                "bodyTruncated": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "folder": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isRead": {
                    "type": "boolean"
                },
                "preview": {
                    "type": "string"
                },
                "receivedAt": {
                    "type": "string"
                },
                "remoteId": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "threadId": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/models.MailErrorAction"
                },
                "code": {
                    "$ref": "#/definitions/models.MailErrorCode"
                },
                "error": {
                    "type": "string"
                },
                "retryAfterSeconds": {
                    "type": "integer"
                }
            }
        },
        "models.Folder": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.MailErrorAction": {
            "type": "string",
            "enum": [
                "none",
                "retry",
                "reconnect",
                "edit_account",
                "contact_admin"
            ],
            "x-enum-varnames": [
                "MailErrorActionNone",
                "MailErrorActionRetry",
                "MailErrorActionReconnect",
                "MailErrorActionEditAccount",
                "MailErrorActionContactAdmin"
            ]
        },
        "models.MailErrorCode": {
            "type": "string",
            "enum": [
                "authentication_failed",
                "permission_denied",
                "organization_policy",
                "quota_exceeded",
                "rate_limited",
                "provider_unavailable",
                "not_found",
                "invalid_configuration",
                "tls_failed",
                "network_failed",
                "protocol_failed",
                "response_too_large",
                "cancelled",
                "internal_error"
            ],
            "x-enum-varnames": [
                "MailErrorAuthentication",
                "MailErrorPermission",
                "MailErrorPolicy",
                "MailErrorQuota",
                "MailErrorRateLimit",
                "MailErrorUnavailable",
                "MailErrorNotFound",
                "MailErrorConfiguration",
                "MailErrorTLS",
                "MailErrorNetwork",
                "MailErrorProtocol",
                "MailErrorResponseTooLarge",
                "MailErrorCancelled",
                "MailErrorInternal"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}